Inline split_biased for slight performance improvement.
Where a generic function uses a small non-generic function, it is
generally a good idea to mark the non-generic function as inline.
This is because generic functions effectively get compiled in a
different module, as do inline functions.  Non-inlined non-generic
functions do not, so are compiled in a different module to their
caller, preventing the compiler from inlining automatically.