Add a benchmark for sequences with short runs
This benchmark demonstrates a strength of gidsort over the standard
sort.

Apart from being good advertising, having a benchmark that
demonstrates the strengths is good for my morale.  It can be very
discouraging to see minor wins for gidsort while being crushed in
the big leagues of randomness.

Also, as the code changes to improve random sorts, this benchmark
helps show if we're compromising in other areas.